user_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
python-dotenv
PyJWT
alembic
python-multipart
pydantic
